Impl linter/toggle
