+++ crates/linter/src/linters/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod go;
+pub(crate) mod rust;
+pub(crate) mod sh;
+pub(crate) mod vim;
