+++ crates/linter/src/linters/sh.rs
